refacto room feature

+++ app/Http/Controllers/RoomsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Room;
+use App\Models\Category;
 use Illuminate\Support\Facades\Log;
 
 class RoomsController extends Controller
@@ -13,26 +14,24 @@ class RoomsController extends Controller
      */
     public function index()
     {
-        $rooms = Room::with('category')->get()->makeHidden('category_id');
+        // $rooms = Room::with('category')->get()->makeHidden('category_id');
+        $rooms = Room::with('category')->get();
+        $categories = Category::all();
 
+        if (!$rooms || !$categories)
+            return response()->json([
+                'message' => 'Rooms or Categories not found.',
+            ], 404);
+        
         return response()->json([
-            'message' => 'Rooms successfully fetched.',
+            'message' => 'Rooms and Categories successfully fetched.',
             // 'title' => 'All Rooms',
             // 'route_name' => 'rooms',
             // 'name' => session()->get('user.name'),
-            'data' => $rooms
-        ]);
+            'rooms' => $rooms,
+            'categories' => $categories
+        ], 200);
     }
-
-    // public function getById(string $id)
-    // {
-    //     $room = Room::find($id);
-
-    //     return response()->json([
-    //         'message' => 'Room successfully fetched.',
-    //         'data' => $room
-    //     ]);
-    // }
     
     /**
      * Store a newly created resource in storage.
@@ -42,7 +41,7 @@ class RoomsController extends Controller
         $request['number'] = (int) $request['number'];
         $request['category_id'] = (int) $request['category_id'];
         $request['price'] = (int) $request['price'];
-                
+        
         $data = $request->validate([
             'number' => 'required|integer',
             'price' => 'required|integer',
@@ -51,7 +50,6 @@ class RoomsController extends Controller
         ]);
 
         Room::create($data);
-
         return response()->json([
             'message' => 'Room successfully created.',
             'data' => $data,
@@ -74,11 +72,17 @@ class RoomsController extends Controller
             'category_id' => 'required|integer',
         ]);
 
-        Room::find($id)->update($data);
+        $room = Room::find($id);
+
+        if (!$room)
+            return response()->json([
+                'message' => 'Room not found.',
+            ], 404);
         
+        $room->update($data);
         return response()->json([
             'message' => 'Room successfully updated.',
-        ]);
+        ], 200);
     }
 
     /**
@@ -86,10 +90,16 @@ class RoomsController extends Controller
      */
     public function destroy(string $id)
     {
+        $room = Room::find($id);
+    
+        if (!$room)
+            return response()->json([
+                'message' => 'Room not found.',
+            ], 404);
+    
         Room::destroy($id);
-
         return response()->json([
             'message' => 'Room successfully deleted.',
-        ]);
+        ], 200);
     }
 }
